requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
